requirements.txt
py-cord
Flask
pymongo
dnspython
youtube-dl
python-dotenv
FFmpeg
craiyon.py
PyNaCl
Pillow